--- server/api/v1/docs/index.js
@@ -1,3 +1,0 @@
-const router = require('./router');
-
-module.exports = { router };
