+++ server/api/v1/docs/router.js
@@ -4,7 +4,7 @@ const swaggerUi = require('swagger-ui-express');
 const { port } = require('../../../../config');
 const { authPaths, authSchemas, authTag } = require('../auth/docs');
 const { healthPaths, healthSchemas, healthTag } = require('../health/docs');
-const { usersSchemas } = require('../users');
+const { usersSchemas } = require('../users/docs');
 
 const openApiDoc = {
   openapi: '3.0.0',
