+++ server/api/v1/index.js
@@ -3,10 +3,10 @@ const express = require('express');
 const router = express.Router();
 
 // Functional endpoints
-router.use('/auth', require('./auth').router);
-router.use('/health', require('./health').router);
+router.use('/auth', require('./auth/router'));
+router.use('/health', require('./health/router'));
 
 // OpenAPI documentation
-router.use('/docs', require('./docs').router);
+router.use('/docs', require('./docs/router'));
 
 module.exports = router;
