--- server/api/v1/users/index.js
@@ -1,7 +1,0 @@
-const userDocs = require('./docs');
-const userEntityElements = require('./entity');
-
-module.exports = {
-  ...userDocs,
-  ...userEntityElements,
-};
